RavenDB-13771 Storage summary on the dashboard doesn't include the system storage
