+++ src/Raven.Server/Web/System/AdminStorageHandler.cs
@@ -1,9 +1,9 @@
 ﻿using System.Threading.Tasks;
-using Raven.Server.Documents;
 using Raven.Server.Routing;
 using Raven.Server.Utils;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
+using Voron;
 
 namespace Raven.Server.Web.System
 {
